2023/pyproject.toml
[tool.poetry]
name = "aoc2022"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
ipython = "^8.18.1"
ruff = "^0.1.6"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
